Lots of work to get Ansible working with Docker
A new role was created to install the circulation manager with
Docker. The old role, which puts it together from Git, is still here
in case we want to switch back for whatever reason.

Additionally, new fields were added to the settings sample file, and
the config template was cleaned up and updated.