A big commit for ECS functionality
I probably could've done this sooner, but this is just about all of
the ECS functionality we need right now. The playbook has been
updated to create an ECS cluster, a task definition for running
the CM, CM scripts, and Elasticsearch Docker containers, and a
service to run the task on any EC2 instance that is part of the
cluster.