Tweaked some settings for stable deployment
A few resource settings needed to be changed to get the deployment
working correctly. A step was also added at the end to reboot the
server, but it's possible this will still need to be done manually.

Additionally, a fully filled out config file has been added to the
repo as a better "sample", with the original sample being renamed
to "template".