Removed content providers from settings file
The problem here is, if certain provider configs are left blank,
they'll still be ingested into the database as of version 1.1.23.
This causes problems in the Circulation Manager when it tries to read
config info from the DB that wasn't present when it was read straight
from the file.

From now on, config info will need to be added to the config.json
template file. If this file isn't modified before running the
playbook, an error will be raised at runtime.