Made some changes to get in line with CM 2.X
The config template has been pared down a bit since a lot of the
configuration now happens through the admin interface. The .yml
template has been updated to be more clear and the sample has been
removed because it wasn't clear enough as it was.

Most importantly, the database URI has been added as an environment
variable for the ECS task.